# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EAMS API",
    "description": "Events Attendance Monitoring System — attendance capture and administration.\n\n**This document supersedes `docs/api/attendance-contract-handoff.md` as the contract.**\nIt is generated from the source, so the outcome tokens, status codes and payload shapes\nbelow are what the running build actually does.\n\n- **Base path** `/api/v1`. JSON in and out, camelCase field names.\n- **All timestamps are UTC**, ISO 8601, with an explicit `Z`. A bare or offset-bearing\n  local time misjudges the Present/Late boundary by the server's offset.\n- **Errors are RFC 7807** `application/problem+json`. Every body carries a `traceId` to\n  quote back and, on the attendance surface, a stable `code` to branch on — never the\n  human-readable `title` or `detail`, which are reworded freely.\n- **Capture endpoints require a device key** (`DeviceKey` below). Everything else is open\n  for now: authentication for human users is Phase 6, and until it lands this API must\n  stay on a local or trusted network.\n\nSee the `TapOutcomeCode`, `ManualOutcomeCode` and `LiveOutcomeCode` schemas for the frozen\ntoken tables.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/devices": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "`GET /devices` — every registered device in this school.",
        "description": "<b>No key material, ever.</b> The plaintext exists only in the 201 that issued it; this returns\r\n            the public key id and the lifecycle timestamps an operator needs to decide whether a credential\r\n            is still in use.",
        "responses": {
          "200": {
            "description": "The devices, possibly empty.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "§6.6 `POST /devices` — register, and issue the first key.",
        "description": "<b>The 201 body is the only time this key is ever readable.</b> The server stores\r\n            `SHA-256(secret)` and nothing else, so there is no \"show it again\" endpoint to write and\r\n            none to forget to protect. An operator who loses it rotates.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceWriteRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceWriteRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceWriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceKeyIssuedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceKeyIssuedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceKeyIssuedDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{id}": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "`GET /devices/{id}` — one registered device.",
        "description": "<b>No key material, ever</b> — same rule as the list. The plaintext key exists only in the 201\r\n            that issued it; what comes back here is the public key id and the lifecycle timestamps.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The device.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The device.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              }
            }
          },
          "404": {
            "description": "No such device in this school.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Devices"
        ],
        "summary": "§6.6 `PUT /devices/{id}` — the device's own fields. Never touches the key; see\r\n`Device.ApiKeyRevokedAt` for why retiring a device and burning its credential are two\r\ndifferent statements.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceWriteRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceWriteRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{id}/regenerate-key": {
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "§6.6 `POST /devices/{id}/regenerate-key` — hard cut, no overlap window. The previous key\r\nstops working the moment this returns.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceKeyIssuedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceKeyIssuedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceKeyIssuedDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{id}/revoke-key": {
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "`POST /devices/{id}/revoke-key` — burn the credential without issuing a replacement.",
        "description": "\r\n\r\n  <b>Not defined by §6.6, and added deliberately.</b> §6.6 offers only regeneration, which\r\n             conflates \"this key is compromised\" with \"give me a working one\" — but the published mobile\r\n             contract already distinguishes the two on the wire (`401` bad key versus `403` revoked\r\n             key), and without this route that 403 is unreachable and `Devices.ApiKeyRevokedAt` is a\r\n             column nothing can write. Recorded as drift.\r\n             \r\n\r\n\r\n             Idempotent: revoking twice, or revoking a device that never held a key, is a 200. The\r\n             postcondition — this device cannot authenticate — holds either way, so a retry after a timeout\r\n             is safe on the one operation an operator runs when something has gone wrong.\r\n             ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{id}/heartbeat": {
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "§6.6 `POST /devices/{id}/heartbeat` — liveness, and one of the four endpoints a device key\r\nauthenticates.",
        "description": "\r\n\r\n  <b>The route id must be the authenticated device's own</b> — the same D-26 rule the tap path\r\n             applies, and it took a second look to see that it belongs here too.\r\n             \r\n\r\n\r\n             The first version of this action skipped the check, on the reasoning that one kiosk stamping a\r\n             sibling's `LastSeenAt` within its own school is a wrong liveness reading rather than a\r\n             data-integrity problem. That undercounts what this writes:\r\n             M:EAMS.Application.Abstractions.IDeviceService.HeartbeatAsync(System.Guid,System.Threading.CancellationToken) also\r\n             sets `ApiKeyLastUsedAt`, and <em>that</em> column is the signal an operator reads when\r\n             deciding whether a credential is still in use and therefore whether it is safe to revoke. A\r\n             retired-but-not-revoked kiosk that a neighbour keeps making look active is a key that never gets\r\n             burned. The check is shorter than the paragraph that argued against it.\r\n             \r\n\r\n\r\n  <b>404, not 403.</b> A device asking about a device that is not itself learns only that the URL\r\n             names nothing it can see — the same non-disclosure rule D-27 applies when it reuses\r\n             `DeviceNotRegistered` rather than inventing a cross-tenant 403.\r\n             ",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "DeviceKey": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "deviceType": {
            "type": "string"
          },
          "readerModel": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "apiKeyId": {
            "type": "string",
            "description": "The <em>public</em> half of the key — the twelve characters between `eams_dk_` and the secret.\r\nSafe to show and to log: it identifies the credential without being one, which is what makes\r\n\"device 3's key id is `a91f…`, and the log line that failed says `a91f…`\" a possible\r\nsentence. Null when the device has never been issued a key.",
            "nullable": true
          },
          "hasActiveKey": {
            "type": "boolean",
            "description": "Whether this device could authenticate right now: it has a key, the key is not revoked, and the\r\ndevice is active. A single boolean rather than three, because \"why can this kiosk not tap?\" is one\r\nquestion and the three columns answer it together."
          },
          "apiKeyIssuedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "apiKeyLastUsedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "apiKeyRevokedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Technical Plan §4.10 / §6.6, as a client sees it.\r\n            \r\n\r\n<b>There is deliberately no key-shaped field on this record.</b> Not the secret, not the hash, not\r\nthe token. The plaintext key exists in exactly two responses — the 201 from `POST /devices` and\r\nthe 200 from `POST /devices/{id}/regenerate-key` — and in EAMS.Application.Dtos.DeviceKeyIssuedDto\r\nalone. Adding a `apiKey` here to \"make the list more useful\" would turn every\r\n`GET /devices` into a credential dump, which is precisely why\r\n`DeviceLifecycleTests.No_read_response_ever_carries_the_key` asserts on the serialized bytes\r\nrather than on the record's shape.\r\n"
      },
...
    },
    "securitySchemes": {
      "DeviceKey": {
        "type": "apiKey",
        "description": "The kiosk / capture credential, scoped to `attendance.capture` and nothing else.\n\nSend the **entire header value including the scheme word**:\n\n```\nAuthorization: DeviceKey eams_dk_<keyId>_<secret>\n```\n\nIssued once by `POST /api/v1/devices` and never retrievable afterwards — the server stores\nonly a hash. Store it in the platform keystore (`expo-secure-store` / Android Keystore),\nnever in `AsyncStorage`. Lose it and rotate with `POST /api/v1/devices/{id}/regenerate-key`.\n\nCase is significant and lower-case is the only accepted form for the token itself.\n\nFailures: `401 DeviceKeyMissing` / `DeviceKeyMalformed` / `DeviceKeyInvalid`,\n`403 DeviceKeyRevoked` / `DeviceInactive`, `429 RateLimited` (honour `Retry-After`).\n`DeviceKeyInvalid` deliberately does not distinguish an unknown key id from a wrong secret.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
...
}